Initial framework
The project is currently in a position where I can begin working on the
core goal of the project. The framework has been setup to allow for the
use of both SFML and OpenGL. I have created a 'Handler' class that
handles the creation of the window and rendering to said window, as well
as any events that come through such as closing and resizing the window.

Next I will begin configuring OpenGL and try to get a texture rendering
to the window.