Rendered triangles
Basic testing has been completed as I have managed to render triangles
to the screen that make use of an VAO, VBO, EBO, and both vertex and
fragment shaders. Next I will be finalising the testing by rendering a
texture to the triangles, from there I can being working on the main
goal of the project.