Texture loading
Textures are now being loaded through SFML's sf::Image class. We can
then generate textures using OpenGL, passing in the data from the stored
sf::Image. This marks the end of the testing I needed to do before I
could begin working on the project. Because of this there won't be as
much programming over the course of the next week as I will now be
working on a writted document that covers the implementation details of
the project.