Texture data
The first iteration of the diffuse, height and normal textures have been
added to the project. These are all being passed into the fragment
shader for later use. I have tested the data being passed to the shader
by outputting it's data to the screen and I can say that it is all
working correctly. The next step is to begin using the data to calculate
the shading of the scene.