Simple height
The lighting shader is now making use of the height map. Currently it is
just set to illuminate each pixel that is below the light. The height of
the light can be changed using the arrow keys. Obviously this will be
changed in the future as point lights should be able to illuminate areas
above them. For now it will be left as is as I still need to test the
normal map to ensure it works as well.