Inevitable mess
This past week has been spent testing the normal map that I had created
in a previous week. I began by just using the standard equation for
calculating diffuse lighting; this proved to work somewhat more than I
was expecting and proved that the normal values were correct. Obviously
the diffuse equation is not something that I want to be using with pixel
art as we want to keep the colour count as low as possible in order to
avoid colour gradients. Additionally the calculations being done are
used for 3D environments so they have some issues when used for 2D.
Because of this I am currently in the process of planning how I am going
to tackle these problems; this should be completed tomorrow after the
scheduled meeting.