Inevitable mess 2
I am currently having troubles using the normals and height values
correctly to determine which pixels should be lit up. I am committing
the current state of the project so that I can talk about it in the
meeting today; however this is in an unfinished state at the moment.