Potential solution
After talking with Keith he seems to have agreed that most of the math
is correct and the issue lies elsewhere. He decided to output the
normalized normal value I was using and the colours being output are
incorrects; up should be green however it was displaying blue etc. Later
I will be working to correct the normal values.