Pixelated light
The shading created by the light is now more pixelated. This is to
better align with the pixel art style. This is simply achieved by using
modulus on the position of the light and the current fragment to round
all numbers to a number divisible by 3. (The art is being scaled up 3
times)

Also I have made some progress in getting the normals and height to
function properly. I have remade the normal map to use colours that
better represent the direction they are facing. I have removed the
scaling function that was scaling our 0 - 1 vectors onto a scale of -1 -
1. This has been replaced by multiplying the values by 2 and then
subtracting 1.