Updating comments
After the shader had gone under some reworking the old comments were
outdated. They have been updated and should be correct. Unfortunately
there has been no progress these past two weeks; there are still a lot
of unexplained issues with the shader that I cannot figure out. I have
run through the shader multiple times to ensure that the math is correct
and it seems as though it should provide better results than it does at
the moment. Issues such as the light being cut in two aren't making much
sense to me and I am not sure how the issue is being produced; making it
much harder to fix.

There is a little over 2 weeks left until the hand-in and I will need to
use that time to write the report for the project and ensure that is up
to the required standard. Because of this I do not expect to make any
more changes to this project as I do not have the time.